+++ lib/kubereq/pod_exec.ex
@@ -86,6 +86,15 @@ defmodule Kubereq.PodExec do
     Kubereq.Connect.start_link(__MODULE__, req, %{into: into})
   end
 
+  def child_spec(init_arg) do
+    default = %{
+      id: __MODULE__,
+      start: {Kubereq.PodExec, :start_link, [init_arg]}
+    }
+
+    Supervisor.child_spec(default, [])
+  end
+
   @doc """
   Check if the websocket is open.
   """